+++ meta/defaults/Workflows.json
@@ -4,7 +4,7 @@
 "versions":{
   "default":{
     "version":"default",
-    "get-method":"download",
+    "get-method":"direct",
     "invoke": "workflows.php",
     "files":{
       "one":"https://raw.githubusercontent.com/jdfwarrior/Workflows/master/workflows.php"
@@ -12,7 +12,7 @@
   },
   "3.3":{
    "version":"3.3",
-   "get-method":"download",
+   "get-method":"direct",
    "invoke": "workflows.php",
    "files":{
     "one":"https://raw.githubusercontent.com/jdfwarrior/Workflows/master/workflows.php"
